put tokenomics utils functions in own file
